Bring ImportJson on par with other methods
